spinlock fix for fsresize
move to compatibility file

organise set-capacity

add compatibility fix for suse

apply compatibility fix
